Dodan front-end, provjera preklapanja s postojećim terminom kodpotvrđivanja termina, provjera postoje li aktivne narudžbe za određenu uslugu

+++ logic/denyActivity.php
@@ -3,7 +3,15 @@
 require_once('../core/init.php');
 
 $activityID = Input::get('activityID');
-Database::getInstance()->query("DELETE FROM Activity WHERE ActivityID=?", array($activityID));
+$count = Database::getInstance()->query("SELECT * FROM Book WHERE ActivityID=?", array($activityID))->count();
 
-echo $activityID;
+if($count > 0){
+    echo "key";
+}
+else{
+    Database::getInstance()->query("DELETE FROM Activity WHERE ActivityID=?", array($activityID));
+    echo $activityID;
+}
+
+
